+++ src/com/doubleyellow/scoreboard/prefs/PreferenceValues.java
@@ -1560,8 +1560,8 @@ public class PreferenceValues extends RWValues
                 // very first install/run
 
                 int appVersionCode = RWValues.getAppVersionCode(context);
-                final int    NO_SHOWCASE_FOR_VERSION        = 205;
-                final String NO_SHOWCASE_FOR_VERSION_BEFORE = "2018-11-25";
+                final int    NO_SHOWCASE_FOR_VERSION        = 206;
+                final String NO_SHOWCASE_FOR_VERSION_BEFORE = "2018-12-18";
                 if ( appVersionCode > NO_SHOWCASE_FOR_VERSION ) {
                     // need to adjust the datecheck below
                     Log.w(TAG, "[getStartupAction] Adjust version code check!!");